# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guest Pro Hotel Concierge API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type LoginRequestType =
  (typeof LoginRequestType)[keyof typeof LoginRequestType];

export const LoginRequestType = {
  manager: "manager",
  guest: "guest",
} as const;

export interface LoginRequest {
  type: LoginRequestType;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  guestKey?: string | null;
}

export type LoginResponseRole =
  (typeof LoginResponseRole)[keyof typeof LoginResponseRole];

export const LoginResponseRole = {
  manager: "manager",
  guest: "guest",
} as const;

export type CurrentUserRole =
  (typeof CurrentUserRole)[keyof typeof CurrentUserRole];

export const CurrentUserRole = {
  manager: "manager",
  guest: "guest",
} as const;

export interface CurrentUser {
  id: number;
  role: CurrentUserRole;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  roomNumber?: string | null;
  /** @nullable */
  guestId?: number | null;
  hotelId: number;
}

export interface LoginResponse {
  role: LoginResponseRole;
  token: string;
  user: CurrentUser;
}

export interface Guest {
  id: number;
  firstName: string;
  lastName: string;
  roomNumber: string;
  hotelId: number;
  createdAt: string;
  /** @nullable */
  guestKey?: string | null;
}

export interface CreateGuestRequest {
  firstName: string;
  lastName: string;
  roomNumber: string;
}

export interface CreateGuestResponse {
  guest: Guest;
  guestKey: string;
}

export type ChatSessionStatus =
  (typeof ChatSessionStatus)[keyof typeof ChatSessionStatus];

export const ChatSessionStatus = {
  active: "active",
  closed: "closed",
  transferred: "transferred",
} as const;

export interface ChatSession {
  id: number;
  guestId: number;
  hotelId: number;
  status: ChatSessionStatus;
  createdAt: string;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
  system: "system",
} as const;

export interface Message {
  id: number;
  sessionId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  originalContent?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  category?: string | null;
  createdAt: string;
}

export interface SendMessageRequest {
  content: string;
  /** @nullable */
  language?: string | null;
}

export interface SendMessageResponse {
  userMessage: Message;
  assistantMessage: Message;
}

export interface HotelBranding {
  hotelId: number;
  appName: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  accentColor?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  welcomeText?: string | null;
}

export interface QuickAction {
  id: number;
  hotelId: number;
  label: string;
  /** @nullable */
  icon?: string | null;
  category: string;
  sortOrder: number;
}