# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guest Pro Hotel Concierge API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login (manager or guest)
 */
export const LoginBody = zod.object({
  type: zod.enum(["manager", "guest"]),
  email: zod.string().nullish(),
  password: zod.string().nullish(),
  guestKey: zod.string().nullish(),
});

export const LoginResponse = zod.object({
  role: zod.enum(["manager", "guest"]),
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    role: zod.enum(["manager", "guest"]),
    email: zod.string().nullish(),
    firstName: zod.string().nullish(),
    lastName: zod.string().nullish(),
    roomNumber: zod.string().nullish(),
    guestId: zod.number().nullish(),
    hotelId: zod.number(),
  }),
});

/**
 * @summary Logout current user
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current session user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  role: zod.enum(["manager", "guest"]),
  email: zod.string().nullish(),
  firstName: zod.string().nullish(),
  lastName: zod.string().nullish(),
  roomNumber: zod.string().nullish(),
  guestId: zod.number().nullish(),
  hotelId: zod.number(),
});

/**
 * @summary List all guests (manager only)
 */
export const ListGuestsResponseItem = zod.object({
  id: zod.number(),
  firstName: zod.string(),
  lastName: zod.string(),
  roomNumber: zod.string(),
  hotelId: zod.number(),
  createdAt: zod.coerce.date(),
  guestKey: zod.string().nullish(),
});
export const ListGuestsResponse = zod.array(ListGuestsResponseItem);

/**
 * @summary Create a new guest (manager only)
 */
export const CreateGuestBody = zod.object({
  firstName: zod.string(),
  lastName: zod.string(),
  roomNumber: zod.string(),
});

/**
 * @summary Get guest by ID
 */
export const GetGuestParams = zod.object({
  id: zod.coerce.number(),
});

export const GetGuestResponse = zod.object({
  id: zod.number(),
  firstName: zod.string(),
  lastName: zod.string(),
  roomNumber: zod.string(),
  hotelId: zod.number(),
  createdAt: zod.coerce.date(),
  guestKey: zod.string().nullish(),
});

/**
 * @summary Create or retrieve active chat session for guest
 */
export const CreateChatSessionResponse = zod.object({
  id: zod.number(),
  guestId: zod.number(),
  hotelId: zod.number(),
  status: zod.enum(["active", "closed", "transferred"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get all messages in a chat session
 */
export const GetChatMessagesParams = zod.object({
  sessionId: zod.coerce.number(),
});

export const GetChatMessagesResponseItem = zod.object({
  id: zod.number(),
  sessionId: zod.number(),
  role: zod.enum(["user", "assistant", "system"]),
  content: zod.string(),
  originalContent: zod.string().nullish(),
  language: zod.string().nullish(),
  category: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetChatMessagesResponse = zod.array(GetChatMessagesResponseItem);

/**
 * @summary Send a message and get AI response
 */
export const SendMessageParams = zod.object({
  sessionId: zod.coerce.number(),
});

export const SendMessageBody = zod.object({
  content: zod.string(),
  language: zod.string().nullish(),
});

export const SendMessageResponse = zod.object({
  userMessage: zod.object({
    id: zod.number(),
    sessionId: zod.number(),
    role: zod.enum(["user", "assistant", "system"]),
    content: zod.string(),
    originalContent: zod.string().nullish(),
    language: zod.string().nullish(),
    category: zod.string().nullish(),
    createdAt: zod.coerce.date(),
  }),
  assistantMessage: zod.object({
    id: zod.number(),
    sessionId: zod.number(),
    role: zod.enum(["user", "assistant", "system"]),
    content: zod.string(),
    originalContent: zod.string().nullish(),
    language: zod.string().nullish(),
    category: zod.string().nullish(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get hotel branding for current tenant
 */
export const GetHotelBrandingResponse = zod.object({
  hotelId: zod.number(),
  appName: zod.string(),
  tagline: zod.string().nullish(),
  primaryColor: zod.string().nullish(),
  accentColor: zod.string().nullish(),
  logoUrl: zod.string().nullish(),
  welcomeText: zod.string().nullish(),
});

/**
 * @summary List quick actions for hotel
 */
export const ListQuickActionsResponseItem = zod.object({
  id: zod.number(),
  hotelId: zod.number(),
  label: zod.string(),
  icon: zod.string().nullish(),
  category: zod.string(),
  sortOrder: zod.number(),
});
export const ListQuickActionsResponse = zod.array(ListQuickActionsResponseItem);