# lib/api-zod/src/generated/types/chatSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guest Pro Hotel Concierge API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatSessionStatus } from "./chatSessionStatus";

export interface ChatSession {
  id: number;
  guestId: number;
  hotelId: number;
  status: ChatSessionStatus;
  createdAt: Date;
}