# lib/api-zod/src/generated/types/createGuestRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guest Pro Hotel Concierge API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGuestRequest {
  firstName: string;
  lastName: string;
  roomNumber: string;
}