# lib/api-zod/src/generated/types/createGuestResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guest Pro Hotel Concierge API
 * OpenAPI spec version: 0.1.0
 */
import type { Guest } from "./guest";

export interface CreateGuestResponse {
  guest: Guest;
  guestKey: string;
}