# lib/api-zod/src/generated/types/hotelBranding.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guest Pro Hotel Concierge API
 * OpenAPI spec version: 0.1.0
 */

export interface HotelBranding {
  hotelId: number;
  appName: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  primaryColor?: string | null;
  /** @nullable */
  accentColor?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  welcomeText?: string | null;
}