# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guest Pro Hotel Concierge API
 * OpenAPI spec version: 0.1.0
 */
import type { LoginRequestType } from "./loginRequestType";

export interface LoginRequest {
  type: LoginRequestType;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  guestKey?: string | null;
}