# lib/api-zod/src/generated/types/quickAction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guest Pro Hotel Concierge API
 * OpenAPI spec version: 0.1.0
 */

export interface QuickAction {
  id: number;
  hotelId: number;
  label: string;
  /** @nullable */
  icon?: string | null;
  category: string;
  sortOrder: number;
}